[planning]: move TrackingErrorCalculator to util.py
