+++ aerial_robot_planning/scripts/pub_mpc_joint_traj.py
@@ -7,8 +7,6 @@ import sys
 import os
 import math
 from abc import ABC, abstractmethod
-
-import numpy as np
 import rospy
 import tf_conversions as tf
 
@@ -16,90 +14,12 @@ from nav_msgs.msg import Odometry
 from geometry_msgs.msg import Transform, Twist, Quaternion, Vector3, Pose
 from trajectory_msgs.msg import MultiDOFJointTrajectory, MultiDOFJointTrajectoryPoint
 
-from util import check_position_initialized
+from util import check_position_initialized, TrackingErrorCalculator
 
 # Insert current folder into path so we can import from "trajs" or other local files
 current_path = os.path.abspath(os.path.dirname(__file__))
 if current_path not in sys.path:
     sys.path.insert(0, current_path)
-
-
-class TrackingErrorCalculator:
-    def __init__(self):
-        self.all_pos_err = [[], [], []]  # x, y, z
-        self.all_ang_err = [[], [], []]  # roll, pitch, yaw
-
-    def reset(self):
-        self.__init__()
-
-    def update(self, odom, traj_msg):
-        err_px, err_py, err_pz, err_roll, err_pitch, err_yaw = self._cal_tracking_error(odom, traj_msg)
-        self.all_pos_err[0].append(err_px)
-        self.all_pos_err[1].append(err_py)
-        self.all_pos_err[2].append(err_pz)
-        self.all_ang_err[0].append(err_roll)
-        self.all_ang_err[1].append(err_pitch)
-        self.all_ang_err[2].append(err_yaw)
-
-        return err_px, err_py, err_pz, err_roll, err_pitch, err_yaw
-
-    def get_rmse_error(self):
-        all_pos_err_np = np.array(self.all_pos_err)
-        all_ang_err_np = np.array(self.all_ang_err)
-
-        pos_rmse = np.sqrt(np.mean(all_pos_err_np ** 2, axis=1))
-        pos_rmse_norm = np.linalg.norm(pos_rmse)
-        ang_rmse = np.sqrt(np.mean(all_ang_err_np ** 2, axis=1))
-        ang_rmse_norm = np.linalg.norm(ang_rmse)
-
-        return pos_rmse_norm, pos_rmse, ang_rmse_norm, ang_rmse
-
-    @staticmethod
-    def _cal_tracking_error(uav_odom: Odometry, ref_traj):
-        """
-        Calculate position and orientation error between current odom and ref_traj,
-        which can be a MultiDOFJointTrajectory or PredXU.
-        """
-        # Current pose
-        cur_pos = uav_odom.pose.pose.position
-        q_cur = [uav_odom.pose.pose.orientation.x,
-                 uav_odom.pose.pose.orientation.y,
-                 uav_odom.pose.pose.orientation.z,
-                 uav_odom.pose.pose.orientation.w]
-
-        # Extract reference pose
-        try:
-            if hasattr(ref_traj, "points"):  # MultiDOFJointTrajectory
-                ref_px = ref_traj.points[0].transforms[0].translation.x
-                ref_py = ref_traj.points[0].transforms[0].translation.y
-                ref_pz = ref_traj.points[0].transforms[0].translation.z
-                ref_qx = ref_traj.points[0].transforms[0].rotation.x
-                ref_qy = ref_traj.points[0].transforms[0].rotation.y
-                ref_qz = ref_traj.points[0].transforms[0].rotation.z
-                ref_qw = ref_traj.points[0].transforms[0].rotation.w
-            else:  # PredXU
-                ref_px = ref_traj.x.data[0]
-                ref_py = ref_traj.x.data[1]
-                ref_pz = ref_traj.x.data[2]
-                ref_qx = ref_traj.x.data[6]
-                ref_qy = ref_traj.x.data[7]
-                ref_qz = ref_traj.x.data[8]
-                ref_qw = ref_traj.x.data[9]
-        except AttributeError:
-            raise AttributeError("Reference trajectory must be either MultiDOFJointTrajectory or PredXU!")
-
-        # Position error
-        dx = cur_pos.x - ref_px
-        dy = cur_pos.y - ref_py
-        dz = cur_pos.z - ref_pz
-
-        # Orientation error
-        m_cur = tf.transformations.quaternion_matrix(q_cur)
-        m_ref = tf.transformations.quaternion_matrix([ref_qx, ref_qy, ref_qz, ref_qw])
-        m_err = m_cur.dot(m_ref.T)
-        euler_err = tf.transformations.euler_from_matrix(m_err, axes="sxyz")
-
-        return dx, dy, dz, euler_err[0], euler_err[1], euler_err[2]
 
 
 ##########################################
